Make alloca SCEV rewriting sound
The AllocaOffsetRewriter is not sound as it's used in
StackSafetyAnalysis. In the SafeStack pass, it's only used to rewrite
alloca expressions so it's safe because the only valid operation (as far
as I'm aware) that you can perform with an address is addition. If you
perform ptrtoint and inttoptr, the value in the SCEV will be the
inttoptr value so the rewriting will be sound.

However, in StackSafetyAnalysis we can actually have the Alloca value be
an integer, e.g. if we're analyzing an integer parameter. If a function
does a ptrtoint, passes the int as a parameter, and the destination
function performs computation with that parameter then it's possible we
will perform unsound 0 rewriting (e.g. for the RHS of a division.)

Limit rewriting to additions only and add assert()s to catch dangerous
empty-set ranges.